LD-2: Updated unsigned ints to use uint8_t.

+++ include/lib/core/die/Die.h
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <stdint.h>
+
 typedef struct Die {
-    unsigned int maxValue;
-    unsigned int currentValue;
+    uint8_t maxValue;
+    uint8_t currentValue;
 } Die;
 
-Die* buildDie(unsigned int maxValue);
+Die* buildDie(uint8_t maxValue);
 void deleteDie(Die** die);
 void rollDie(Die* die);
